split up the tests
